+++ okhttp/src/main/java/com/squareup/okhttp/internal/http/SpdyTransport.java
@@ -20,8 +20,8 @@ import com.squareup.okhttp.Headers;
 import com.squareup.okhttp.Protocol;
 import com.squareup.okhttp.Request;
 import com.squareup.okhttp.Response;
-import com.squareup.okhttp.internal.bytes.ByteString;
 import com.squareup.okhttp.internal.Util;
+import com.squareup.okhttp.internal.bytes.ByteString;
 import com.squareup.okhttp.internal.spdy.ErrorCode;
 import com.squareup.okhttp.internal.spdy.Header;
 import com.squareup.okhttp.internal.spdy.SpdyConnection;
@@ -187,6 +187,10 @@ public final class SpdyTransport implements Transport {
         .headers(headersBuilder.build());
   }
 
+  @Override public void emptyTransferStream() {
+    // Do nothing.
+  }
+
   @Override public InputStream getTransferStream(CacheRequest cacheRequest) throws IOException {
     return new SpdyInputStream(stream, cacheRequest, httpEngine);
   }
@@ -224,6 +228,90 @@ public final class SpdyTransport implements Transport {
       throw new AssertionError(protocol);
     }
     return prohibited;
+  }
+
+  /**
+   * An input stream for the body of an HTTP response.
+   *
+   * <p>Since a single socket's input stream may be used to read multiple HTTP
+   * responses from the same server, subclasses shouldn't close the socket stream.
+   *
+   * <p>A side effect of reading an HTTP response is that the response cache
+   * is populated. If the stream is closed early, that cache entry will be
+   * invalidated.
+   */
+  abstract static class AbstractHttpInputStream extends InputStream {
+    protected final InputStream in;
+    protected final HttpEngine httpEngine;
+    private final CacheRequest cacheRequest;
+    protected final OutputStream cacheBody;
+    protected boolean closed;
+
+    AbstractHttpInputStream(InputStream in, HttpEngine httpEngine, CacheRequest cacheRequest)
+        throws IOException {
+      this.in = in;
+      this.httpEngine = httpEngine;
+
+      OutputStream cacheBody = cacheRequest != null ? cacheRequest.getBody() : null;
+
+      // Some apps return a null body; for compatibility we treat that like a null cache request.
+      if (cacheBody == null) {
+        cacheRequest = null;
+      }
+
+      this.cacheBody = cacheBody;
+      this.cacheRequest = cacheRequest;
+    }
+
+    /**
+     * read() is implemented using read(byte[], int, int) so subclasses only
+     * need to override the latter.
+     */
+    @Override public final int read() throws IOException {
+      return Util.readSingleByte(this);
+    }
+
+    protected final void checkNotClosed() throws IOException {
+      if (closed) {
+        throw new IOException("stream closed");
+      }
+    }
+
+    protected final void cacheWrite(byte[] buffer, int offset, int count) throws IOException {
+      if (cacheBody != null) {
+        cacheBody.write(buffer, offset, count);
+      }
+    }
+
+    /**
+     * Closes the cache entry and makes the socket available for reuse. This
+     * should be invoked when the end of the body has been reached.
+     */
+    protected final void endOfInput() throws IOException {
+      if (cacheRequest != null) {
+        cacheBody.close();
+      }
+      httpEngine.release(false);
+    }
+
+    /**
+     * Calls abort on the cache entry and disconnects the socket. This
+     * should be invoked when the connection is closed unexpectedly to
+     * invalidate the cache entry and to prevent the HTTP connection from
+     * being reused. HTTP messages are sent in serial so whenever a message
+     * cannot be read to completion, subsequent messages cannot be read
+     * either and the connection must be discarded.
+     *
+     * <p>An earlier implementation skipped the remaining bytes, but this
+     * requires that the entire transfer be completed. If the intention was
+     * to cancel the transfer, closing the connection is the only solution.
+     */
+    protected final void unexpectedEndOfInput() {
+      if (cacheRequest != null) {
+        cacheRequest.abort();
+      }
+      httpEngine.release(true);
+    }
   }
 
   /** An HTTP message body terminated by the end of the underlying stream. */
